# Improving printing performance