--- a/src/binary-operator-printers/exponentiation.js
+++ b/src/binary-operator-printers/exponentiation.js
@@ -5,7 +5,7 @@ const { group, indent, line } = doc.builders;
 export const exponentiation = {
   match: (op) => op === '**',
   print: (node, path, print) => {
-    const right = [' ', node.operator, line, path.call(print, 'right')];
+    const right = [` ${node.operator}`, line, path.call(print, 'right')];
     // If it's a single binary operation, avoid having a small right
     // operand like - 1 on its own line
     const shouldGroup =