--- a/src/nodes/AssemblyLocalDefinition.js
+++ b/src/nodes/AssemblyLocalDefinition.js
@@ -4,17 +4,9 @@ import { printSeparatedList } from '../common/printer-helpers.js';
 const { line } = doc.builders;
 
 export const AssemblyLocalDefinition = {
-  print: ({ node, path, print }) => {
-    const parts = [
-      'let',
-      printSeparatedList(path.map(print, 'names'), { firstSeparator: line })
-    ];
-
-    if (node.expression !== null) {
-      parts.push(':= ');
-      parts.push(path.call(print, 'expression'));
-    }
-
-    return parts;
-  }
+  print: ({ node, path, print }) => [
+    'let',
+    printSeparatedList(path.map(print, 'names'), { firstSeparator: line }),
+    node.expression ? [':= ', path.call(print, 'expression')] : ''
+  ]
 };