--- a/src/nodes/AssemblyStackAssignment.js
+++ b/src/nodes/AssemblyStackAssignment.js
@@ -1,7 +1,6 @@
 export const AssemblyStackAssignment = {
   print: ({ node, path, print }) => [
     path.call(print, 'expression'),
-    ' =: ',
-    node.name
+    ` =: ${node.name}`
   ]
 };