--- a/src/nodes/ElementaryTypeName.js
+++ b/src/nodes/ElementaryTypeName.js
@@ -1,5 +1,5 @@
 const stateMutability = (node) =>
-  node.stateMutability?.length > 0 ? [' ', node.stateMutability] : '';
+  node.stateMutability?.length > 0 ? ` ${node.stateMutability}` : '';
 
 export const ElementaryTypeName = {
   print: ({ node }) => [node.name, stateMutability(node)]