--- a/src/nodes/EnumDefinition.js
+++ b/src/nodes/EnumDefinition.js
@@ -1,17 +1,16 @@
 import { doc } from 'prettier';
 import { printSeparatedList } from '../common/printer-helpers.js';
 
-const { group, hardline } = doc.builders;
+const { hardline } = doc.builders;
 
 export const EnumDefinition = {
-  print: ({ node, path, print }) =>
-    group([
-      'enum ',
-      node.name,
-      ' {',
-      printSeparatedList(path.map(print, 'members'), {
-        firstSeparator: hardline
-      }),
-      '}'
-    ])
+  print: ({ node, path, print }) => [
+    'enum ',
+    node.name,
+    ' {',
+    printSeparatedList(path.map(print, 'members'), {
+      firstSeparator: hardline
+    }),
+    '}'
+  ]
 };