--- a/src/nodes/EventDefinition.js
+++ b/src/nodes/EventDefinition.js
@@ -7,12 +7,8 @@ const parameters = (node, path, print) =>
 
 export const EventDefinition = {
   print: ({ node, path, print }) => [
-    'event ',
-    node.name,
-    '(',
+    `event ${node.name}(`,
     parameters(node, path, print),
-    ')',
-    node.isAnonymous ? ' anonymous' : '',
-    ';'
+    `)${node.isAnonymous ? ' anonymous' : ''};`
   ]
 };