--- a/src/nodes/ExpressionStatement.js
+++ b/src/nodes/ExpressionStatement.js
@@ -5,23 +5,15 @@ const { hardline } = doc.builders;
 
 export const ExpressionStatement = {
   print: ({ node, options, path, print }) => {
-    const parts = [];
+    const comments =
+      path.getParentNode().type === 'IfStatement'
+        ? printComments(node, path, options)
+        : [];
 
-    const parent = path.getParentNode();
-
-    if (parent.type === 'IfStatement') {
-      if (node.comments?.length) {
-        const comments = printComments(node, path, options);
-        if (comments?.length) {
-          parts.push(comments);
-          parts.push(hardline);
-        }
-      }
-    }
-
-    parts.push(path.call(print, 'expression'));
-    parts.push(node.omitSemicolon ? [] : ';');
-
-    return parts;
+    return [
+      comments.length ? [comments, hardline] : '',
+      path.call(print, 'expression'),
+      node.omitSemicolon ? '' : ';'
+    ];
   }
 };