--- a/src/nodes/IfStatement.js
+++ b/src/nodes/IfStatement.js
@@ -6,26 +6,25 @@ import {
 
 const { group, hardline, indent, line } = doc.builders;
 
-const printTrueBody = (node, path, print) => {
-  if (node.trueBody.type === 'Block') {
-    return [' ', path.call(print, 'trueBody')];
-  }
-
-  return group(indent([line, path.call(print, 'trueBody')]), {
-    shouldBreak: node.trueBody.type === 'IfStatement' // `if` within `if`
-  });
-};
+const printTrueBody = (node, path, print) =>
+  node.trueBody.type === 'Block'
+    ? [' ', path.call(print, 'trueBody')]
+    : group(indent([line, path.call(print, 'trueBody')]), {
+        shouldBreak: node.trueBody.type === 'IfStatement' // `if` within `if`
+      });
 
 const printFalseBody = (node, path, print) =>
   node.falseBody.type === 'Block' || node.falseBody.type === 'IfStatement'
     ? [' ', path.call(print, 'falseBody')]
     : group(indent([line, path.call(print, 'falseBody')]));
 
-const printElse = (node, path, print, commentsBetweenIfAndElse) => {
+const printElse = (node, path, print, options) => {
   if (node.falseBody) {
+    const comments = printComments(node, path, options);
     return [
-      node.trueBody.type !== 'Block' || commentsBetweenIfAndElse.length > 0 // `else` on new line
-        ? hardline
+      comments.length ? [hardline, comments] : '',
+      node.trueBody.type !== 'Block' || comments.length > 0
+        ? hardline // `else` on new line
         : ' ',
       'else',
       printFalseBody(node, path, print)
@@ -35,22 +34,11 @@ const printElse = (node, path, print, commentsBetweenIfAndElse) => {
 };
 
 export const IfStatement = {
-  print: ({ node, options, path, print }) => {
-    const comments = node.comments || [];
-    const commentsBetweenIfAndElse = comments.filter(
-      (comment) => !comment.leading && !comment.trailing
-    );
-
-    const parts = [];
-
-    parts.push('if (', printSeparatedItem(path.call(print, 'condition')), ')');
-    parts.push(printTrueBody(node, path, print));
-    if (commentsBetweenIfAndElse.length && node.falseBody) {
-      parts.push(hardline);
-      parts.push(printComments(node, path, options));
-    }
-    parts.push(printElse(node, path, print, commentsBetweenIfAndElse));
-
-    return parts;
-  }
+  print: ({ node, options, path, print }) => [
+    'if (',
+    printSeparatedItem(path.call(print, 'condition')),
+    ')',
+    printTrueBody(node, path, print),
+    printElse(node, path, print, options)
+  ]
 };