--- a/src/nodes/LabelDefinition.js
+++ b/src/nodes/LabelDefinition.js
@@ -3,5 +3,5 @@ import { doc } from 'prettier';
 const { dedent, line } = doc.builders;
 
 export const LabelDefinition = {
-  print: ({ node }) => [dedent(line), node.name, ':']
+  print: ({ node }) => [dedent(line), `${node.name}:`]
 };