--- a/src/nodes/ModifierDefinition.js
+++ b/src/nodes/ModifierDefinition.js
@@ -5,16 +5,15 @@ const { group, hardline, indent, line } = doc.builders;
 
 const modifierParameters = (node, path, print) => {
   if (node.parameters?.length > 0) {
+    // To keep consistency any list of parameters will split if it's longer than 2.
+    // For more information see:
+    // https://github.com/prettier-solidity/prettier-plugin-solidity/issues/256
+    const shouldBreak = node.parameters.length > 2;
     return [
       '(',
       printSeparatedList(path.map(print, 'parameters'), {
-        separator: [
-          ',',
-          // To keep consistency any list of parameters will split if it's longer than 2.
-          // For more information see:
-          // https://github.com/prettier-solidity/prettier-plugin-solidity/issues/256
-          node.parameters.length > 2 ? hardline : line
-        ]
+        separator: [',', shouldBreak ? hardline : line],
+        grouped: !shouldBreak
       }),
       ')'
     ];