--- a/src/nodes/NumberLiteral.js
+++ b/src/nodes/NumberLiteral.js
@@ -1,6 +1,6 @@
 export const NumberLiteral = {
   print: ({ node }) =>
     node.subdenomination
-      ? [node.number, ' ', node.subdenomination]
+      ? `${node.number} ${node.subdenomination}`
       : node.number
 };