--- a/src/nodes/PragmaDirective.js
+++ b/src/nodes/PragmaDirective.js
@@ -1,3 +1,3 @@
 export const PragmaDirective = {
-  print: ({ node }) => ['pragma ', node.name, ' ', node.value, ';']
+  print: ({ node }) => `pragma ${node.name} ${node.value};`
 };