--- a/src/nodes/StringLiteral.js
+++ b/src/nodes/StringLiteral.js
@@ -9,7 +9,7 @@ export const StringLiteral = {
       (part, index) =>
         // node.isUnicode is an array of the same length as node.parts
         // that indicates if that string fragment has the unicode prefix
-        (node.isUnicode[index] ? 'unicode' : '') + printString(part, options)
+        `${node.isUnicode[index] ? 'unicode' : ''}${printString(part, options)}`
     );
 
     return join(hardline, list);