--- a/src/nodes/TupleExpression.js
+++ b/src/nodes/TupleExpression.js
@@ -1,7 +1,4 @@
-import { doc } from 'prettier';
 import { printSeparatedList } from '../common/printer-helpers.js';
-
-const { group } = doc.builders;
 
 const contents = (node, path, print) =>
   node.components?.length === 1 && node.components[0].type === 'BinaryOperation'
@@ -9,10 +6,9 @@ const contents = (node, path, print) =>
     : printSeparatedList(path.map(print, 'components'));
 
 export const TupleExpression = {
-  print: ({ node, path, print }) =>
-    group([
-      node.isArray ? '[' : '(',
-      contents(node, path, print),
-      node.isArray ? ']' : ')'
-    ])
+  print: ({ node, path, print }) => [
+    node.isArray ? '[' : '(',
+    contents(node, path, print),
+    node.isArray ? ']' : ')'
+  ]
 };