--- a/src/nodes/TypeDefinition.js
+++ b/src/nodes/TypeDefinition.js
@@ -1,3 +1,3 @@
 export const TypeDefinition = {
-  print: ({ node }) => ['type ', node.name, ' is ', node.definition.name, ';']
+  print: ({ node }) => `type ${node.name} is ${node.definition.name};`
 };