--- a/src/nodes/UnaryOperation.js
+++ b/src/nodes/UnaryOperation.js
@@ -2,7 +2,7 @@ export const UnaryOperation = {
   print: ({ node, path, print }) =>
     node.isPrefix
       ? [
-          node.operator + (node.operator === 'delete' ? ' ' : ''),
+          `${node.operator}${node.operator === 'delete' ? ' ' : ''}`,
           path.call(print, 'subExpression')
         ]
       : [path.call(print, 'subExpression'), node.operator]