--- a/src/nodes/UncheckedStatement.js
+++ b/src/nodes/UncheckedStatement.js
@@ -1,7 +1,3 @@
-import { doc } from 'prettier';
-
-const { group } = doc.builders;
-
 export const UncheckedStatement = {
-  print: ({ path, print }) => group(['unchecked ', path.call(print, 'block')])
+  print: ({ path, print }) => ['unchecked ', path.call(print, 'block')]
 };