--- a/src/nodes/UsingForDeclaration.js
+++ b/src/nodes/UsingForDeclaration.js
@@ -12,7 +12,7 @@ export const UsingForDeclaration = {
           printSeparatedList(
             node.functions.map((functionName, i) =>
               node.operators[i]
-                ? [functionName, ' as ', node.operators[i]]
+                ? `${functionName} as ${node.operators[i]}`
                 : functionName
             ),
             {