Removing `group`s that will always break:

- UncheckedStatement
- EnumDefinition
- ModifierDefinition

Removing `group` that just affect a closing bracket `)` or `]` that will be outside the `tabWidth`

- TupleExpression

Member Access Chain is already properly `group`ed

- MemberAccess

Just reordering the `group` `indent` calls:

- FunctionDefinition
- FunctionTypeName

Prefer using `shouldBreak` at the `group` level instead of choosing between `hardline` and `line`
- IfStatement

The rest is wrapping strings together whenever possible.
Finally I rebased some functions that would `push` and `concat` into a variable and returned an built array instead.