@Janther 2024-01-11
Thank you for this,
after this [comment](https://github.com/prettier-solidity/prettier-plugin-solidity/pull/953#issuecomment-1868340445) addressing the same issue I decided to keep the original behaviour although in the future if prettier decides to adopt `experimentalTernaries` as the default formatting choice and we decide to follow that choice we will need to sometimes add parentheses the thing will be when to remove them.

But that's for the future.