@Janther 2024-01-11
All of the changes of this PR will be included in the  TypeScript #968 step by step migration so I decided to close this one in favor to that one